creates a busted blender scene :D
